Other: Introduce AlignmentUI plugin.
